Simplify the MS365 sample
