+++ how-to/integrate-with-ms365/client/src/apps.ts
@@ -1,158 +1,55 @@
+import type OpenFin from "@openfin/core";
 import type { App } from "@openfin/workspace";
-import { getSettings } from "./settings";
+import { AppManifestType, getCurrentSync } from "@openfin/workspace-platform";
 
-let cachedApps: App[];
-
-async function validateEntries(apps: App[]) {
-	let canLaunchExternalProcessResponse;
-
-	try {
-		canLaunchExternalProcessResponse = await fin.System.queryPermissionForCurrentContext(
-			"System.launchExternalProcess"
-		);
-	} catch (error) {
-		console.error("Error while querying for System.launchExternalProcess permission", error);
-	}
-	const canLaunchExternalProcess = canLaunchExternalProcessResponse?.granted;
-
-	let canDownloadAppAssetsResponse;
-
-	try {
-		canDownloadAppAssetsResponse = await fin.System.queryPermissionForCurrentContext("System.downloadAsset");
-	} catch (error) {
-		console.error("Error while querying for System.downloadAsset permission", error);
+/**
+ * Launch the passed app using its manifest type to determine how to launch it.
+ * @param app The app to launch.
+ * @returns The value returned by the launch.
+ */
+export async function launchApp(
+	app: App
+): Promise<OpenFin.Platform | OpenFin.Identity | OpenFin.View | OpenFin.Application | undefined> {
+	if (!app.manifest) {
+		console.error(`No manifest was provided for type ${app.manifestType}`);
+		return;
 	}
 
-	const canDownloadAppAssets = canDownloadAppAssetsResponse?.granted;
+	let ret: OpenFin.Platform | OpenFin.Identity | OpenFin.View | OpenFin.Application | undefined;
 
-	if (canLaunchExternalProcess && canDownloadAppAssets) {
-		return apps;
-	}
+	console.log("Application launch requested:", app);
 
-	const validatedApps: App[] = [];
-	const rejectedAppIds = [];
-	const settings = await getSettings();
-	const appAssetTag: string = settings?.appProvider?.appAssetTag ?? "appasset";
+	switch (app.manifestType) {
+		case AppManifestType.Snapshot: {
+			const platform = getCurrentSync();
+			ret = await platform.applySnapshot(app.manifest);
+			break;
+		}
 
-	for (let i = 0; i < apps.length; i++) {
-		if (apps[i].manifestType !== "external") {
-			validatedApps.push(apps[i]);
-		} else if (!canLaunchExternalProcess) {
-			rejectedAppIds.push(apps[i].appId);
-		} else if (Array.isArray(apps[i].tags) && apps[i].tags.includes(appAssetTag) && !canDownloadAppAssets) {
-			rejectedAppIds.push(apps[i].appId);
-		} else {
-			validatedApps.push(apps[i]);
+		case AppManifestType.View: {
+			const platform = getCurrentSync();
+			ret = await platform.createView({ manifestUrl: app.manifest });
+			break;
+		}
+
+		case AppManifestType.External: {
+			ret = await fin.System.launchExternalProcess({ path: app.manifest, uuid: app.appId });
+			break;
+		}
+
+		// We have a custom manifest type which will open the msteams: link
+		case "desktop-browser": {
+			await fin.System.openUrlWithBrowser(app.manifest);
+			break;
+		}
+
+		default: {
+			ret = await fin.Application.startFromManifest(app.manifest);
+			break;
 		}
 	}
 
-	if (rejectedAppIds.length > 0) {
-		console.warn(
-			"Apps.ts: validateEntries: Not passing the following list of applications as they will not be able to run on this machine due to missing permissions. Alternatively this logic could be moved to the launch function where a selection is not launched but the user is presented with a modal saying they can't launch it due to permissions.",
-			rejectedAppIds
-		);
-	}
+	console.log("Finished application launch request");
 
-	return validatedApps;
-}
-
-async function getRestEntries(
-	url: string,
-	credentials?: "omit" | "same-origin" | "include",
-	cacheDuration?: number
-): Promise<App[]> {
-	const options = credentials !== undefined ? { credentials } : undefined;
-	if (url === undefined) {
-		return [];
-	}
-	const response = await fetch(url, options);
-
-	const apps: App[] = await response.json();
-
-	cachedApps = await validateEntries(apps);
-
-	if (cacheDuration !== undefined) {
-		const setTimeoutInMs = cacheDuration * 60 * 1000;
-		setTimeout(() => {
-			console.log("Clearing cache of apps as cache duration has passed.");
-			cachedApps = undefined;
-		}, setTimeoutInMs);
-	}
-
-	return cachedApps;
-}
-
-export async function getApps(): Promise<App[]> {
-	console.log("Requesting apps.");
-	try {
-		const settings = await getSettings();
-		const apps =
-			cachedApps ??
-			(await getRestEntries(
-				settings?.appProvider?.appsSourceUrl,
-				settings?.appProvider?.includeCredentialOnSourceRequest,
-				settings?.appProvider?.cacheDurationInMinutes
-			));
-		return apps;
-	} catch (err) {
-		console.error("Error retrieving apps. Returning empty list.", err);
-		return [];
-	}
-}
-
-export async function getAppsByTag(tags: string[], mustMatchAll = false): Promise<App[]> {
-	const apps = await getApps();
-	const filteredApps = apps.filter((value) => {
-		if (value.tags === undefined) {
-			return false;
-		}
-		let matchFound = false;
-		for (let i = 0; i < tags.length; i++) {
-			if (value.tags.includes(tags[i])) {
-				if (mustMatchAll) {
-					matchFound = true;
-				} else {
-					return true;
-				}
-			} else if (mustMatchAll) {
-				return false;
-			}
-		}
-		return matchFound;
-	});
-	return filteredApps;
-}
-
-export async function getApp(requestedApp: string | { appId: string }): Promise<App> {
-	const apps = await getApps();
-	let appId;
-	if (requestedApp !== undefined) {
-		if (typeof requestedApp === "string") {
-			appId = requestedApp;
-		} else {
-			appId = requestedApp.appId;
-		}
-	}
-	if (appId === undefined) {
-		return undefined;
-	}
-	const app = apps.find((entry) => entry.appId === appId);
-
-	return app;
-}
-
-export async function getAppsByIntent(intent: string): Promise<App[]> {
-	const apps = await getApps();
-	const filteredApps = apps.filter((value) => {
-		if (value.intents === undefined) {
-			return false;
-		}
-		for (let i = 0; i < value.intents.length; i++) {
-			if (value.intents[i].name.toLowerCase() === intent.toLowerCase()) {
-				return true;
-			}
-		}
-		return false;
-	});
-	return filteredApps;
+	return ret;
 }
