--- how-to/integrate-with-ms365/client/src/default-palletes.ts
@@ -1,63 +1,0 @@
-import type { CustomPaletteSet } from "@openfin/workspace/common/src/api/theming";
-
-export const DEFAULT_PALETTES: {
-	dark: CustomPaletteSet;
-	light: CustomPaletteSet;
-} = {
-	light: {
-		brandPrimary: "#0A76D3",
-		brandSecondary: "#1E1F23",
-		backgroundPrimary: "#FAFBFE",
-		background1: "#FFFFFF",
-		background2: "#FAFBFE",
-		background3: "#F3F5F8",
-		background4: "#ECEEF1",
-		background5: "#DDDFE4",
-		background6: "#C9CBD2",
-		statusSuccess: "#35C759",
-		statusWarning: "#F48F00",
-		statusCritical: "#BE1D1F",
-		statusActive: "#0498FB",
-		inputBackground: "#ECEEF1",
-		inputColor: "#1E1F23",
-		inputPlaceholder: "#383A40",
-		inputDisabled: "#7D808A",
-		inputFocused: "#C9CBD2",
-		textDefault: "#1E1F23",
-		textHelp: "#2F3136",
-		textInactive: "#7D808A",
-		contentBackground1: "#0A76D3",
-		contentBackground2: "#000000",
-		contentBackground3: "#000000",
-		contentBackground4: "#000000",
-		contentBackground5: "#000000"
-	},
-	dark: {
-		brandPrimary: "#0A76D3",
-		brandSecondary: "#383A40",
-		backgroundPrimary: "#1E1F23",
-		background1: "#111214",
-		background2: "#1E1F23",
-		background3: "#24262B",
-		background4: "#2F3136",
-		background5: "#383A40",
-		background6: "#53565F",
-		statusSuccess: "#35C759",
-		statusWarning: "#F48F00",
-		statusCritical: "#BE1D1F",
-		statusActive: "#0498FB",
-		inputBackground: "#53565F",
-		inputColor: "#FFFFFF",
-		inputPlaceholder: "#C9CBD2",
-		inputDisabled: "#7D808A",
-		inputFocused: "#C9CBD2",
-		textDefault: "#FFFFFF",
-		textHelp: "#C9CBD2",
-		textInactive: "#7D808A",
-		contentBackground1: "#0A76D3",
-		contentBackground2: "#000000",
-		contentBackground3: "#000000",
-		contentBackground4: "#000000",
-		contentBackground5: "#000000"
-	}
-};
