--- how-to/integrate-with-ms365/client/src/integrations/ms365/index.ts
@@ -1,5 +1,0 @@
-import { Microsoft365Provider } from "./integration-provider";
-
-export const entryPoints: { [id: string]: Microsoft365Provider } = {
-	integrations: new Microsoft365Provider()
-};
