+++ how-to/integrate-with-ms365/client/src/module-entry-point.ts
@@ -1,0 +1,9 @@
+import { Microsoft365Integration } from "./ms365-integration";
+
+/**
+ * This entry point is used by webpack to create a module file which can be consumed
+ * by the customize-workspace example.
+ */
+export const entryPoints: { [id: string]: Microsoft365Integration } = {
+	integrations: new Microsoft365Integration()
+};
