--- how-to/integrate-with-ms365/client/src/shapes/logger-shapes.ts
@@ -1,38 +1,0 @@
-export interface Logger {
-	/**
-	 * Log data as information.
-	 * @param message The message to log.
-	 * @param optionalParams Optional parameters for details.
-	 */
-	info(message: unknown, ...optionalParams: unknown[]): void;
-
-	/**
-	 * Log data as error.
-	 * @param message The message to log.
-	 * @param optionalParams Optional parameters for details.
-	 */
-	error(message: unknown, ...optionalParams: unknown[]): void;
-
-	/**
-	 * Log data as warning.
-	 * @param message The message to log.
-	 * @param optionalParams Optional parameters for details.
-	 */
-	warn(message: unknown, ...optionalParams: unknown[]): void;
-
-	/**
-	 * Log data as trace.
-	 * @param message The message to log.
-	 * @param optionalParams Optional parameters for details.
-	 */
-	trace(message: unknown, ...optionalParams: unknown[]): void;
-
-	/**
-	 * Log data as debug.
-	 * @param message The message to log.
-	 * @param optionalParams Optional parameters for details.
-	 */
-	debug(message: unknown, ...optionalParams: unknown[]): void;
-}
-
-export type LoggerCreator = (group: string) => Logger;
