--- how-to/integrate-with-ms365/client/src/shapes/theme-shapes.ts
@@ -1,4 +1,0 @@
-export enum ColorSchemeMode {
-	Light = "light",
-	Dark = "dark"
-}
