--- how-to/integrate-with-ms365/public/apps.json
@@ -1,192 +1,0 @@
-[
-	{
-		"appId": "ms-365-teams",
-		"name": "ms-365-teams",
-		"title": "Microsoft Teams",
-		"description": "Launch Microsoft Teams.",
-		"manifest": "msteams:",
-		"manifestType": "desktop-browser",
-		"icons": [
-			{
-				"src": "http://localhost:8080/images/apps/teams.svg"
-			}
-		],
-		"contactEmail": "contact@example.com",
-		"supportEmail": "support@example.com",
-		"publisher": "OpenFin",
-		"intents": [],
-		"images": [],
-		"tags": ["view", "microsoft", "365", "teams"]
-	},
-	{
-		"appId": "ms-365-outlook-mail",
-		"name": "ms-365-outlook-mail",
-		"title": "Microsoft Outlook Mail",
-		"description": "Launch Microsoft 365 Outlook Mail in a view.",
-		"manifest": "http://localhost:8080/views/ms-365-outlook-mail-view.json",
-		"manifestType": "view",
-		"icons": [
-			{
-				"src": "http://localhost:8080/images/apps/outlook-mail.svg"
-			}
-		],
-		"contactEmail": "contact@example.com",
-		"supportEmail": "support@example.com",
-		"publisher": "OpenFin",
-		"intents": [],
-		"images": [],
-		"tags": ["view", "microsoft", "365", "outlook", "mail"]
-	},
-	{
-		"appId": "ms-365-outlook-calendar",
-		"name": "ms-365-outlook-calendar",
-		"title": "Microsoft Outlook Calendar",
-		"description": "Launch Microsoft 365 Outlook Calendar in a view.",
-		"manifest": "http://localhost:8080/views/ms-365-outlook-calendar-view.json",
-		"manifestType": "view",
-		"icons": [
-			{
-				"src": "http://localhost:8080/images/apps/outlook-calendar.svg"
-			}
-		],
-		"contactEmail": "contact@example.com",
-		"supportEmail": "support@example.com",
-		"publisher": "OpenFin",
-		"intents": [],
-		"images": [],
-		"tags": ["view", "microsoft", "365", "outlook", "calendar"]
-	},
-	{
-		"appId": "ms-365-outlook-people",
-		"name": "ms-365-outlook-people",
-		"title": "Microsoft Outlook People",
-		"description": "Launch Microsoft 365 Outlook People in a view.",
-		"manifest": "http://localhost:8080/views/ms-365-outlook-people-view.json",
-		"manifestType": "view",
-		"icons": [
-			{
-				"src": "http://localhost:8080/images/apps/outlook-people.svg"
-			}
-		],
-		"contactEmail": "contact@example.com",
-		"supportEmail": "support@example.com",
-		"publisher": "OpenFin",
-		"intents": [],
-		"images": [],
-		"tags": ["view", "microsoft", "365", "outlook", "people"]
-	},
-	{
-		"appId": "ms-365-word",
-		"name": "ms-365-word",
-		"title": "Microsoft Word",
-		"description": "Launch Microsoft 365 Word in a view.",
-		"manifest": "http://localhost:8080/views/ms-365-word-view.json",
-		"manifestType": "view",
-		"icons": [
-			{
-				"src": "http://localhost:8080/images/apps/word.svg"
-			}
-		],
-		"contactEmail": "contact@example.com",
-		"supportEmail": "support@example.com",
-		"publisher": "OpenFin",
-		"intents": [],
-		"images": [],
-		"tags": ["view", "microsoft", "365", "word"]
-	},
-	{
-		"appId": "ms-365-excel",
-		"name": "ms-365-excel",
-		"title": "Microsoft Excel",
-		"description": "Launch Microsoft 365 Excel in a view.",
-		"manifest": "http://localhost:8080/views/ms-365-excel-view.json",
-		"manifestType": "view",
-		"icons": [
-			{
-				"src": "http://localhost:8080/images/apps/excel.svg"
-			}
-		],
-		"contactEmail": "contact@example.com",
-		"supportEmail": "support@example.com",
-		"publisher": "OpenFin",
-		"intents": [],
-		"images": [],
-		"tags": ["view", "microsoft", "365", "excel"]
-	},
-	{
-		"appId": "ms-365-powerpoint",
-		"name": "ms-365-powerpoint",
-		"title": "Microsoft PowerPoint",
-		"description": "Launch Microsoft 365 PowerPoint in a view.",
-		"manifest": "http://localhost:8080/views/ms-365-powerpoint-view.json",
-		"manifestType": "view",
-		"icons": [
-			{
-				"src": "http://localhost:8080/images/apps/powerpoint.svg"
-			}
-		],
-		"contactEmail": "contact@example.com",
-		"supportEmail": "support@example.com",
-		"publisher": "OpenFin",
-		"intents": [],
-		"images": [],
-		"tags": ["view", "microsoft", "365", "powerpoint"]
-	},
-	{
-		"appId": "ms-365-sharepoint",
-		"name": "ms-365-sharepoint",
-		"title": "Microsoft SharePoint",
-		"description": "Launch Microsoft 365 SharePoint in a view.",
-		"manifest": "http://localhost:8080/views/ms-365-sharepoint-view.json",
-		"manifestType": "view",
-		"icons": [
-			{
-				"src": "http://localhost:8080/images/apps/sharepoint.svg"
-			}
-		],
-		"contactEmail": "contact@example.com",
-		"supportEmail": "support@example.com",
-		"publisher": "OpenFin",
-		"intents": [],
-		"images": [],
-		"tags": ["view", "microsoft", "365", "sharepoint"]
-	},
-	{
-		"appId": "ms-365-onedrive",
-		"name": "ms-365-onedrive",
-		"title": "Microsoft OneDrive",
-		"description": "Launch Microsoft 365 OneDrive in a view.",
-		"manifest": "http://localhost:8080/views/ms-365-onedrive-view.json",
-		"manifestType": "view",
-		"icons": [
-			{
-				"src": "http://localhost:8080/images/apps/onedrive.svg"
-			}
-		],
-		"contactEmail": "contact@example.com",
-		"supportEmail": "support@example.com",
-		"publisher": "OpenFin",
-		"intents": [],
-		"images": [],
-		"tags": ["view", "microsoft", "365", "onedrive"]
-	},
-	{
-		"appId": "ms-365-onenote",
-		"name": "ms-365-onenote",
-		"title": "Microsoft OneNote",
-		"description": "Launch Microsoft 365 OneNote in a view.",
-		"manifest": "http://localhost:8080/views/ms-365-onenote-view.json",
-		"manifestType": "view",
-		"icons": [
-			{
-				"src": "http://localhost:8080/images/apps/onenote.svg"
-			}
-		],
-		"contactEmail": "contact@example.com",
-		"supportEmail": "support@example.com",
-		"publisher": "OpenFin",
-		"intents": [],
-		"images": [],
-		"tags": ["view", "microsoft", "365", "onenote"]
-	}
-]
