+++ how-to/integrate-with-ms365/server/tsconfig.json
@@ -6,6 +6,7 @@
 		"rootDir": "./src",
 		"outDir": "build",
 		"skipLibCheck": true,
+		"strict": true,
 		"esModuleInterop": true,
 		"forceConsistentCasingInFileNames": true,
 		"moduleResolution": "node"
