+++ how-to/integrate-with-sso/integrate-with-okta/client/src/provider.ts
@@ -87,7 +87,7 @@ async function initializeWorkspaceComponents(): Promise<void> {
  * Read the custom settings from the manifest.fin.json.
  * @returns The custom settings from the manifest.
  */
-export async function getManifestCustomSettings(): Promise<CustomSettings> {
+async function getManifestCustomSettings(): Promise<CustomSettings> {
 	// Get the manifest for the current application
 	const app = await fin.Application.getCurrent();
