+++ how-to/register-with-platform-windows/client/src/provider.ts
@@ -47,7 +47,7 @@ async function initializeWorkspacePlatform(): Promise<void> {
 /**
  * Bring the platform to life.
  */
-export async function initializeWorkspaceComponents(): Promise<void> {
+async function initializeWorkspaceComponents(): Promise<void> {
 	console.log("Initialising the bootstrapper");
 
 	await Home.register({
