+++ how-to/register-with-store-basic/client/src/apps.ts
@@ -3,14 +3,6 @@ import type { App } from "@openfin/workspace";
 import { AppManifestType, getCurrentSync } from "@openfin/workspace-platform";
 
 /**
- * Get the list of apps to display.
- * @returns List of app definitions.
- */
-export function getApps(): App[] {
-	return [OPENFIN_INFORMATION_APP, EXPERO_APP, PROCESS_MANAGER, DEVELOPER_CONTENT];
-}
-
-/**
  * App definition to use for demonstration which show OpenFin environment information.
  */
 export const OPENFIN_INFORMATION_APP: App = {
