+++ how-to/use-notifications/client/src/provider.ts
@@ -850,7 +850,7 @@ function addConnectionChangedEventListener(callback: (status: provider.ProviderS
  * Polyfills randomUUID if running in a non-secure context.
  * @returns The random UUID.
  */
-export function randomUUID(): string {
+function randomUUID(): string {
 	if ("randomUUID" in window.crypto) {
 		// eslint-disable-next-line no-restricted-syntax
 		return window.crypto.randomUUID();
